Adding 'm' (bomb) to the toggles: '[o', ']o', 'yo'
I may not use this a lot -- but since the letter is currently free,
it is somewhat worth it.   I spend lots of time developing on Windows
with a lot of other developers using different tools, some of which add
a BOM automatically while others do not.

And to make matters worse... we have some automated scripts that
automatically update the top of the file, which fails to honor the
BOM and sometimes it is no longer the first couple of bytes in the
file :(.

I can use the BOM toggle when repairing this situation.